+++ gems/smithy/lib/smithy/forge/client.rb
@@ -28,7 +28,7 @@ module Smithy
       def source_files
         Enumerator.new do |e|
           e.yield "#{@gem_name}.gemspec", Views::Client::Gemspec.new(@plan).hammer
-          e.yield '.rubocop.yml', Views::Client::Rubocop.new(@plan).hammer
+          e.yield '.rubocop.yml', Views::Client::RubocopYml.new(@plan).hammer
 
           e.yield "lib/#{@gem_name}.rb", Views::Client::Module.new(@plan).hammer
           e.yield "lib/#{@gem_name}/client.rb", Views::Client::Client.new(@plan).hammer
@@ -38,11 +38,10 @@ module Smithy
           e.yield "lib/#{@gem_name}/errors.rb", Views::Client::Errors.new(@plan).hammer
           e.yield "lib/#{@gem_name}/endpoint_parameters.rb", Views::Client::EndpointParameters.new(@plan).hammer
           e.yield "lib/#{@gem_name}/endpoint_provider.rb", Views::Client::EndpointProvider.new(@plan).hammer
-          e.yield "lib/#{@gem_name}/plugins/endpoint.rb", Views::Client::Plugins::Endpoint.new(@plan).hammer
+          e.yield "lib/#{@gem_name}/plugins/endpoint.rb", Views::Client::EndpointPlugin.new(@plan).hammer
 
-          e.yield 'spec/spec_helper.rb', Views::Client::Specs::SpecHelper.new(@plan).hammer
-          e.yield "spec/#{@gem_name}/endpoint_provider_spec.rb",
-                  Views::Client::Specs::EndpointProviderSpec.new(@plan).hammer
+          e.yield 'spec/spec_helper.rb', Views::Client::SpecHelper.new(@plan).hammer
+          e.yield "spec/#{@gem_name}/endpoint_provider_spec.rb", Views::Client::EndpointProviderSpec.new(@plan).hammer
         end
       end
       # rubocop:enable Metrics/AbcSize
