+++ gems/smithy/lib/smithy/forge/types.rb
@@ -25,7 +25,7 @@ module Smithy
       def source_files
         Enumerator.new do |e|
           e.yield "#{@gem_name}.gemspec", Views::Client::Gemspec.new(@plan).hammer
-          e.yield '.rubocop.yml', Views::Client::Rubocop.new(@plan).hammer
+          e.yield '.rubocop.yml', Views::Client::RubocopYml.new(@plan).hammer
           e.yield "lib/#{@gem_name}.rb", Views::Client::Module.new(@plan).hammer
           e.yield "lib/#{@gem_name}/types.rb", Views::Client::Types.new(@plan).hammer
         end
