+++ gems/smithy/lib/smithy/templates/client/endpoint_plugin.erb
