+++ gems/smithy/lib/smithy/templates/client/endpoint_provider_spec.erb
