+++ gems/smithy/lib/smithy/templates/client/rubocop_yml.erb
