+++ gems/smithy/lib/smithy/templates/client/spec_helper.erb
