+++ gems/smithy/lib/smithy/views/client.rb
@@ -18,13 +18,13 @@ require_relative 'client/request_response_example'
 require_relative 'client/client'
 require_relative 'client/customizations'
 require_relative 'client/endpoint_parameters'
+require_relative 'client/endpoint_plugin'
 require_relative 'client/endpoint_provider'
 require_relative 'client/errors'
 require_relative 'client/gemspec'
 require_relative 'client/module'
-require_relative 'client/plugins/endpoint'
-require_relative 'client/rubocop'
+require_relative 'client/rubocop_yml'
 require_relative 'client/shapes'
-require_relative 'client/specs/endpoint_provider_spec'
-require_relative 'client/specs/spec_helper'
+require_relative 'client/endpoint_provider_spec'
+require_relative 'client/spec_helper'
 require_relative 'client/types'
